requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
bs4==0.0.1
pathlib==1.0.1
pytest==5.4.3
pytest-assume==2.2.1
pytest-datadir==1.3.1
pytest-datafiles==2.0
pytest-forked==1.3.0
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-rerunfailures==9.0
pytest-sugar==0.9.4
pytest-xdist==1.34.0
PyYAML==5.3.1
requests==2.24.0


common~=0.1.2
urllib3~=1.25.9
testdata~=1.4.6
beautifulsoup4~=4.9.1